Account for Across minDeposit when bridging funds (#68)
This PR adds logic to account for the `minDeposit` (e.g. min bridge amount) when using the Across bridge. If there are not enough funds on a chain to meet the `minDeposit` amount, the QuarkBuilder will not attempt to bridge from that chain.

We also update the logic for `getTotalAvailableBalance` to calculate the total available balance only on the destination chain. This primarily affects the amount that is used for max actions (e.g. transfer max).