Add executionType to Action (#83)
Adds `executionType` to `Action`. Up until now, this was populated by the app. However, as we dive into a composable actions world with more complex actions, it makes more sense to have this logic live in the QuarkBuilder.

I also updated the acceptance tests to now assert the `executionType` for different operations.