Swap and supply (#87)
This PR implements the first composable action: swap + supply. 

The action intent is currently defined as follows (only Comet for now, Morpho will come in a follow-up PR):
```
struct SwapAndSupplyIntent {
    ZeroExSwapIntent swapIntent;
    CometSupplyIntent supplyIntent;
    // TODO: Add other supply (e.g. morpho) intents
    uint256 blockTimestamp;
    bool preferAcross;
    string paymentAssetSymbol;
}
```

The intent is designed to combine multiple action intents, reducing the number of separate struct definitions to make it easier for clients to call. An alternative approach would be to define a new generalized `SupplyIntent` that has a field to describe the target market (e.g. Comet, Morpho, Aave). However, it should be noted that supply intents across different market types are not uniform (e.g. `MorphoVaultSupplyIntent` does not take a market address while `CometSupplyIntent` does).