Support max supply in QuarkBuilder (#93)
We use the max supply functionality introduced in PR #91. This allows us to pass `type(uint256).max` as the amount to the Quark scripts for supply, which supplies the entire on-chain balance of an account.

Now that we use max supply, we need to re-order where QuotePay sits in a `Multicall` operation. The QuotePay needs to occur *before* the max supply to ensure there are funds left to pay with. We also can't just insert the QuotePay at the very front of the list because it should occur *after* actions that bring in tokens, such as withdraw and swap.