Rough idea of TStoracle
~~This patch adds a simple idea for TStoracle, currently untested. We require EIP-712
signatures for messages and only accept a single signer for a single message with an
expiry. This generally seems like a simple and reasonable approach.~~

We switch to a TStoracle that utilizes non-overwritable keys to achieve a similar purpose
and mostly allow the quark operation to handle signature verifications.

Patches:
  * Continue adding more tests to TStoracle