Capped max transfers (#104)
We add a `cappedMax` flag to the transfer scripts to ensure that "max transfers" are capped at some amount. This protects against the case where a user initiates a max transfer, but receives some funds before that max transfer takes place, causing them to transfer out the recently received funds as well.